requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.1
Flask-Limiter>=3.8
SQLAlchemy>=2.0
alembic>=1.13
Pillow>=10.4
